requirements.txt
matplotlib==3.9.1
Pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
